Creating interface manifest handler file
