+++ packages/office-addin-manifest/src/validate.ts
@@ -3,7 +3,7 @@
 
 import { createReadStream } from "fs";
 import fetch from "node-fetch";
-import { OfficeAddinManifest } from "./manifestInfo";
+import { OfficeAddinManifest } from "./manifestHandler/manifestHandlerXml";
 import { usageDataObject } from "./defaults";
 
 export class ManifestValidationDetails {
